feat(controller): Personalizando validación y status response.

+++ api-rest/src/main/java/com/jeam/api_rest/controller/ClientController.java
@@ -1,8 +1,14 @@
 package com.jeam.api_rest.controller;
 
+
+import java.util.HashMap;
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -45,9 +51,24 @@ public class ClientController {
 		return clientService.save(client);
 	}
 	
-	@DeleteMapping("/cliente/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void deleteClient(@PathVariable Long id) {
-		clientService.delete(id);
+	@DeleteMapping("/cliente/{id}")	
+	public ResponseEntity<?> deleteClient(@PathVariable Long id) {
+		Map<String, Object> response = new HashMap<>();
+		try {
+			Cliente  cliente = clientService.findById(id);	
+			if (cliente == null) {
+				response.put("mensaje",
+						"El cliente con el id ".concat(id.toString().concat(" no existe en la base de datos")));
+				return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			
+            clientService.delete(id);
+            
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (DataAccessException ex) {
+        	response.put("mensaje", "Error al eliminar el cliente en la base de datos");
+            response.put("error", ex.getMessage().concat(": ").concat(ex.getMostSpecificCause().getMessage()));
+            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        }            
 	} 
 }
